add: - change synapse_contract to support swap contract - TODO: add swap contract addresses - change amm code to get contract address from synapse_contract address
