+++ src/common/utils.ts
@@ -95,3 +95,19 @@ const CHAINID_CONTRACTS_MAP: {[c: number]: SynapseContracts.SynapseContract} = {
 }
 
 export const contractsForChainId = (chainId: number): SynapseContracts.SynapseContract => CHAINID_CONTRACTS_MAP[chainId] ?? null
+
+export function getTimeMinutesFromNow(minutesFromNow) {
+    const currentTimeSeconds = new Date().getTime() / 1000
+
+    return Math.round(
+    currentTimeSeconds + 60 * minutesFromNow
+    )
+}
+
+export function uiToNative(amount: number, decimals: number): BigNumber {
+    return BigNumber.from(Math.round(amount * Math.pow(10, decimals)));
+}
+
+export function nativeToUi(amount: number, decimals: number): number {
+    return amount / Math.pow(10, decimals);
+}
